==> code/Downstream_Analysis_Figures_Forrister_et_al_2022/3_Figure_2_normalized_phytochemical_diversity_per_species.Rmd <==
---
title: "Figure_2 Normalized phytochemical_diversity per species"
author: "Dale Forrister"
date: "2/02/2020"
output: html_document
---

```{r setup, include=FALSE}
if (!require(here)) install.packages('here')
library(here)
if (!require(data.table)) install.packages('data.table')
library(data.table)
if (!require(tidyverse)) install.packages('tidyverse')
library(tidyverse)
if (!require(ggplot2)) install.packages('ggplot2')
library(ggplot2)
if (!require(ggpubr)) install.packages('ggpubr')
library(ggpubr)
if (!require(hillR)) install.packages('hillR')
library(hillR)



```
In this paper we use Functional Hill numbers from (Chao, A., Chiu, C.-H. & Jost, L. Annu. Rev. Ecol. Evol. Syst. 45, 297–324, 2014) to calculate the number of equally abundant and equally functionally diverse compounds produced by each species.

To do so we need two types of data:
1) sampsByCompounds = matrix contain expression of each compound in each sample. 
2) pairwise.comps.all_dist = pairwise matrix with cosine disimilarities between compounds. Cosine calculations are produced by GNPS and then converted to distances (1-similarity)

We read these in for the actual data:
```{r}
source(here("./code/0_matching_sampsBycomps_and_pairwise.R"))
```
For the null model we read in all 98 null_sampsByCompounds generated:
```{r}
# files <- list.files(here("./results/null_model/1_SampsByComps_Null_Model/"),pattern = ".csv",full.names = T)
# all_data <- data.frame()
# 
# for(file in files){
#   file_ind <- read.csv(file)
#   file_ind$iteration <- gsub(".csv","",gsub("/Users/dlforrister/Documents_Mac/CODE_GIT_HUB_2017_Aug_31/Evolution_Of_Inga_Chemistry/32_Github_upload_new_phylogney/./results/null_model/1_SampsByComps_Null_Model//Evolved_Random_no_cutoff_","",file))
#   all_data <- rbind(all_data,file_ind)
# }
# 
# names(all_data)[-c(1,which(names(all_data) == "iteration"))] <- gsub("X","",names(all_data)[-c(1,which(names(all_data) == "iteration"))])
# 
# names(all_data)[!names(all_data) %in% names(pairwise.comps.all_dist)]
# all_data[1:10,1:10]

```

The R Packagre HillR calculates all functional numbers for a given community.

The following code will do this. It is very slow so we provide the output in:
./results/phytochemical_diversity_actual_data.csv
./results/null_model/Hill_func_diversity_Evolved_Random_all_null.csv

```{r}
library(RMySQL)
library(doParallel)
library(foreach)
# fields <- c("species_code","itteration", "Raos_Q" , "D_q","MD_q", "FD_q", "ncomps")
# 
# comp_dist_matrix_1 <- as.dist(pairwise.comps.all_dist,upper = T)
# 
# lapply(dbListConnections( dbDriver( drv = "MySQL")), dbDisconnect)
# 
# cores <- detectCores()
# cl <- parallel::makeCluster(8, setup_strategy = "sequential")
# clusterEvalQ(cl,{
#   library(RMySQL)
#   mydb = dbConnect(MySQL(), user= 'u6000251',password = 'sunfun11!',dbname = 'inga_2015_06_01',host='mysql.chpc.utah.edu')
#   NULL
# })
# 
# 
# registerDoParallel(cl)
# 
# #dim(all_data[n,-c(1,which(names(all_data) == "iteration"))])
# 
# 
# foreach(n = 1:nrow(all_data),.packages = "hillR", .errorhandling = "stop") %dopar% {
#   Phyto_diversity <- as.data.frame(t(hill_func(comm = all_data[n,-c(1,which(names(all_data) == "iteration"))], traits = comp_dist_matrix_1, traits_as_is = T,q = 0,fdis = F)))
# 
#   hill_phyto_diversity_random<- data.frame(species_code = as.character(all_data$node_label[n]),itteration=all_data$iteration[n],Raos_Q = Phyto_diversity$Q, D_q = Phyto_diversity$D_q,MD_q = Phyto_diversity$MD_q, FD_q= Phyto_diversity$FD_q, ncomps = rowSums(all_data[n,-c(1,which(names(all_data) == "iteration"))]>0))
# 
#   dbWriteTable(mydb,"A_Hill_func_diversity_Evolved_Random_Q0",hill_phyto_diversity_random, field.types=field,row.names=F,overwrite=F, append=T)
# }

#Pull all of the results from the database and save as a single file:
mydb = dbConnect(MySQL(), user= 'u6000251',password = 'sunfun11!',dbname = 'inga_2015_06_01',host='mysql.chpc.utah.edu')
select <- paste("SELECT * FROM `A_Hill_func_diversity_Evolved_Random_Q0`")

processed_Q0 <- as.data.frame(dbGetQuery(mydb,select))[2:7]
write.csv(processed_Q0,here("./results/null_model/Hill_func_diversity_Evolved_Random_all_null_Q0.csv"),row.names=F)

hill_phyto_diversity_evolved_random_Q0 <- read.csv(here("./results/null_model/Hill_func_diversity_Evolved_Random_all_null_Q0.csv"))
 
boxplot(hill_phyto_diversity_evolved_random_Q0$D_q~hill_phyto_diversity_evolved_random_Q0$itteration, ylab = "Functional Hill Number", xlab = "Null Model Itteration")

dim(hill_phyto_diversity_evolved_random_Q0)

select <- paste("SELECT * FROM `A_Hill_func_diversity_Evolved_Random_Q1`")

processed_Q1 <- as.data.frame(dbGetQuery(mydb,select))[2:7]
write.csv(processed_Q1,here("./results/null_model/Hill_func_diversity_Evolved_Random_all_null_Q1.csv"),row.names=F)

hill_phyto_diversity_evolved_random_Q1 <- read.csv(here("./results/null_model/Hill_func_diversity_Evolved_Random_all_null_Q1.csv"))

dim(hill_phyto_diversity_evolved_random_Q1)

boxplot(hill_phyto_diversity_evolved_random_Q1$D_q~hill_phyto_diversity_evolved_random_Q1$itteration, ylab = "Functional Hill Number", xlab = "Null Model Itteration")

select <- paste("SELECT * FROM `A_Hill_func_diversity_Evolved_Random`")

processed <- as.data.frame(dbGetQuery(mydb,select))[2:7]
write.csv(processed,here("./results/null_model/Hill_func_diversity_Evolved_Random_all_null.csv"),row.names=F)

hill_phyto_diversity_evolved_random_Q2 <- read.csv(here("./results/null_model/Hill_func_diversity_Evolved_Random_all_null.csv"))
 
boxplot(hill_phyto_diversity_evolved_random_Q2$D_q~hill_phyto_diversity_evolved_random_Q2$itteration, ylab = "Functional Hill Number", xlab = "Null Model Itteration")


```
All iterations are very similar,

next we calculate phytochemical diversity for the actual data.
```{r}
comp_dist_matrix_1 <- as.dist(pairwise.comps.all_dist,upper = T)

Phyto_diversity_Q0 <- as.data.frame(t(hill_func(comm = sampsByCompounds, traits = comp_dist_matrix_1, traits_as_is = T,q = 0,fdis = F)))

hill_phyto_diversity_actual_Q0 <- data.frame(species_code = row.names(Phyto_diversity_Q0),Raos_Q = Phyto_diversity_Q0$Q, D_q = Phyto_diversity_Q0$D_q,MD_q = Phyto_diversity_Q0$MD_q, FD_q= Phyto_diversity_Q0$FD_q, ncomps = rowSums(sampsByCompounds>0))

Phyto_diversity_Q1 <- as.data.frame(t(hill_func(comm = sampsByCompounds, traits = comp_dist_matrix_1, traits_as_is = T,q = 1,fdis = F)))

hill_phyto_diversity_actual_Q1 <- data.frame(species_code = row.names(Phyto_diversity_Q1),Raos_Q = Phyto_diversity_Q1$Q, D_q = Phyto_diversity_Q1$D_q,MD_q = Phyto_diversity_Q1$MD_q, FD_q= Phyto_diversity_Q1$FD_q, ncomps = rowSums(sampsByCompounds>0))

Phyto_diversity_Q2 <- as.data.frame(t(hill_func(comm = sampsByCompounds, traits = comp_dist_matrix_1, traits_as_is = T,q = 2,fdis = F)))

hill_phyto_diversity_actual_Q2 <- data.frame(species_code = row.names(Phyto_diversity_Q2),Raos_Q = Phyto_diversity_Q2$Q, D_q = Phyto_diversity_Q2$D_q,MD_q = Phyto_diversity_Q2$MD_q, FD_q= Phyto_diversity_Q2$FD_q, ncomps = rowSums(sampsByCompounds>0))


```

Calculate average for Q=2
```{r}
hill_phyto_diversity_evolved_random_avg_Q2 <- setDT(hill_phyto_diversity_evolved_random_Q2)[,.(Raos_Q_mean = mean(Raos_Q), Raos_Q_sd = sd(Raos_Q), D_q_mean = mean(D_q), D_q_sd =sd(D_q), MD_q_mean = mean(MD_q), MD_q_sd = sd(MD_q)),by="species_code"]

#note that the merge gets rid of the internal nodes that were calculated during in the null model.
hill_phyto_diversity_comparison_Q2 <- merge(hill_phyto_diversity_actual_Q2,hill_phyto_diversity_evolved_random_avg_Q2, by = "species_code") 

hill_phyto_diversity_comparison_Q2$Roas_Q_norm <- (hill_phyto_diversity_comparison_Q2$Raos_Q-hill_phyto_diversity_comparison_Q2$Raos_Q_mean)/hill_phyto_diversity_comparison_Q2$Raos_Q_sd

hill_phyto_diversity_comparison_Q2$MD_q_norm <- (hill_phyto_diversity_comparison_Q2$MD_q-hill_phyto_diversity_comparison_Q2$MD_q_mean)/hill_phyto_diversity_comparison_Q2$MD_q_sd


hill_phyto_diversity_comparison_Q2$D_q_norm <- (hill_phyto_diversity_comparison_Q2$D_q-hill_phyto_diversity_comparison_Q2$D_q_mean)/hill_phyto_diversity_comparison_Q2$D_q_sd



funct_div_for_boxplot_Q2 <- data.frame(species_code = rep(hill_phyto_diversity_comparison_Q2$species_code,2),type = c(rep("Actual_Data",nrow(hill_phyto_diversity_comparison_Q2)),rep("Modeled_Random",nrow(hill_phyto_diversity_comparison_Q2))),Raos_Q = c(hill_phyto_diversity_comparison_Q2$Raos_Q,hill_phyto_diversity_comparison_Q2$Raos_Q_mean), D_q = c(hill_phyto_diversity_comparison_Q2$D_q,hill_phyto_diversity_comparison_Q2$D_q_mean), MD_q = c(hill_phyto_diversity_comparison_Q2$MD_q,hill_phyto_diversity_comparison_Q2$MD_q_mean))



my_comparisons_Q2 <- list( c("Actual_Data", "Modeled_Random"))
p_Q2 <- ggboxplot(funct_div_for_boxplot_Q2, x = "type", y = "D_q",,add = "jitter") + stat_compare_means(comparisons = my_comparisons_Q2, method="t.test",label = "p.signif",label.y = 200) +
  xlab("") + ylab("Phytochemical Diversity\n (Functional Hill Number)") + stat_compare_means(label.y = 200,method="t.test")
 

p_Q2
 
 
```
Calculate average for Q=1
```{r}
hill_phyto_diversity_evolved_random_avg_Q1 <- setDT(hill_phyto_diversity_evolved_random_Q1)[,.(Raos_Q_mean = mean(Raos_Q), Raos_Q_sd = sd(Raos_Q), D_q_mean = mean(D_q), D_q_sd =sd(D_q), MD_q_mean = mean(MD_q), MD_q_sd = sd(MD_q)),by="species_code"]

#note that the merge gets rid of the internal nodes that were calculated during in the null model.
hill_phyto_diversity_comparison_Q1 <- merge(hill_phyto_diversity_actual_Q1,hill_phyto_diversity_evolved_random_avg_Q1, by = "species_code") 

hill_phyto_diversity_comparison_Q1$Roas_Q_norm <- (hill_phyto_diversity_comparison_Q1$Raos_Q-hill_phyto_diversity_comparison_Q1$Raos_Q_mean)/hill_phyto_diversity_comparison_Q1$Raos_Q_sd

hill_phyto_diversity_comparison_Q1$MD_q_norm <- (hill_phyto_diversity_comparison_Q1$MD_q-hill_phyto_diversity_comparison_Q1$MD_q_mean)/hill_phyto_diversity_comparison_Q1$MD_q_sd


hill_phyto_diversity_comparison_Q1$D_q_norm <- (hill_phyto_diversity_comparison_Q1$D_q-hill_phyto_diversity_comparison_Q1$D_q_mean)/hill_phyto_diversity_comparison_Q1$D_q_sd



funct_div_for_boxplot_Q1 <- data.frame(species_code = rep(hill_phyto_diversity_comparison_Q1$species_code,2),type = c(rep("Actual_Data",nrow(hill_phyto_diversity_comparison_Q1)),rep("Modeled_Random",nrow(hill_phyto_diversity_comparison_Q1))),Raos_Q = c(hill_phyto_diversity_comparison_Q1$Raos_Q,hill_phyto_diversity_comparison_Q1$Raos_Q_mean), D_q = c(hill_phyto_diversity_comparison_Q1$D_q,hill_phyto_diversity_comparison_Q1$D_q_mean), MD_q = c(hill_phyto_diversity_comparison_Q1$MD_q,hill_phyto_diversity_comparison_Q1$MD_q_mean))




my_comparisons_Q1 <- list( c("Actual_Data", "Modeled_Random"))
p_Q1 <- ggboxplot(funct_div_for_boxplot_Q1, x = "type", y = "D_q",,add = "jitter") + stat_compare_means(comparisons = my_comparisons_Q1, method="t.test",label = "p.signif",label.y = 200) +
  xlab("") + ylab("Phytochemical Diversity\n (Functional Hill Number)") + stat_compare_means(label.y = 200,method="t.test")
 

p_Q1
 
 
```

Calculate average for Q=0
```{r}
hill_phyto_diversity_evolved_random_avg_Q0 <- setDT(hill_phyto_diversity_evolved_random_Q0)[,.(Raos_Q_mean = mean(Raos_Q), Raos_Q_sd = sd(Raos_Q), D_q_mean = mean(D_q), D_q_sd =sd(D_q), MD_q_mean = mean(MD_q), MD_q_sd = sd(MD_q)),by="species_code"]

#note that the merge gets rid of the internal nodes that were calculated during in the null model.
hill_phyto_diversity_comparison_Q0 <- merge(hill_phyto_diversity_actual_Q0,hill_phyto_diversity_evolved_random_avg_Q0, by = "species_code") 

hill_phyto_diversity_comparison_Q0$Roas_Q_norm <- (hill_phyto_diversity_comparison_Q0$Raos_Q-hill_phyto_diversity_comparison_Q0$Raos_Q_mean)/hill_phyto_diversity_comparison_Q0$Raos_Q_sd

hill_phyto_diversity_comparison_Q0$MD_q_norm <- (hill_phyto_diversity_comparison_Q0$MD_q-hill_phyto_diversity_comparison_Q0$MD_q_mean)/hill_phyto_diversity_comparison_Q0$MD_q_sd


hill_phyto_diversity_comparison_Q0$D_q_norm <- (hill_phyto_diversity_comparison_Q0$D_q-hill_phyto_diversity_comparison_Q0$D_q_mean)/hill_phyto_diversity_comparison_Q0$D_q_sd



funct_div_for_boxplot_Q0 <- data.frame(species_code = rep(hill_phyto_diversity_comparison_Q0$species_code,2),type = c(rep("Actual_Data",nrow(hill_phyto_diversity_comparison_Q0)),rep("Modeled_Random",nrow(hill_phyto_diversity_comparison_Q0))),Raos_Q = c(hill_phyto_diversity_comparison_Q0$Raos_Q,hill_phyto_diversity_comparison_Q0$Raos_Q_mean), D_q = c(hill_phyto_diversity_comparison_Q0$D_q,hill_phyto_diversity_comparison_Q0$D_q_mean), MD_q = c(hill_phyto_diversity_comparison_Q0$MD_q,hill_phyto_diversity_comparison_Q0$MD_q_mean))




my_comparisons_Q0 <- list( c("Actual_Data", "Modeled_Random"))
p_Q0 <- ggboxplot(funct_div_for_boxplot_Q0, x = "type", y = "D_q",,add = "jitter") + stat_compare_means(comparisons = my_comparisons_Q0, method="t.test",label = "p.signif",label.y = 200) +
  xlab("") + ylab("Phytochemical Diversity\n (Functional Hill Number)") + stat_compare_means(label.y = 200,method="t.test")
 

p_Q0
 
 
```

#Compounds in the Actual Dataset are more disimilar to each other (more phytochemically diverse) than expected by random chance

Figure 1B - bargraph Q=2

```{r}

hill_phyto_diversity_comparison_Q2$sig <- "Random"
hill_phyto_diversity_comparison_Q2$sig[which(hill_phyto_diversity_comparison_Q2$D_q_norm <= -2)] <- "Underdispersed"
hill_phyto_diversity_comparison_Q2$sig[which(hill_phyto_diversity_comparison_Q2$D_q_norm >= 2)] <- "Overdisprsed"

#table()
hill_phyto_diversity_comparison_Q2 <- hill_phyto_diversity_comparison_Q2[order(hill_phyto_diversity_comparison_Q2$D_q_norm),]
hill_phyto_diversity_comparison_Q2$order <- 1:nrow(hill_phyto_diversity_comparison_Q2)

phyt_div_plot_Q2 <- ggbarplot(hill_phyto_diversity_comparison_Q2, x = "order", y = "D_q_norm",
          fill = "sig",               # change fill color by cyl
          color = "sig",            # Set bar border colors to white
          palette = "jco",            # jco journal color palett. see ?ggpar
          sort.val = "asc",          # Sort the value in dscending order
          sort.by.groups = FALSE,     # Don't sort inside each group
          x.text.angle = 90           # Rotate vertically x axis texts
          ) + xlab("") + ylab("Normalized Phytochemical Diversity") + theme(axis.title.x=element_blank(),legend.title = element_blank(),
        axis.text.x=element_blank(),axis.ticks.x=element_blank()) + scale_fill_manual(values = c("Underdispersed" = "red", "Random" = "Yellow","Overdisprsed" = "Blue")) + geom_hline(yintercept=2, linetype="dashed", color = "red") + geom_hline(yintercept= -2, linetype="dashed", color = "red")
        

ggsave(phyt_div_plot_Q2,path = here("./results/"),filename = "Figure_2_Phyt_div_plot_D_q_Q2.pdf",dpi = 600,width = 12)
table(hill_phyto_diversity_comparison_Q2$sig)/nrow(hill_phyto_diversity_comparison_Q2)
#Q2 = 38 percent overdispered and 61 percent random

phyt_div_plot_Q2



```
Figure 1B - bargraph Q=1

```{r}

hill_phyto_diversity_comparison_Q1$sig <- "Random"
hill_phyto_diversity_comparison_Q1$sig[which(hill_phyto_diversity_comparison_Q1$D_q_norm <= -2)] <- "Underdispersed"
hill_phyto_diversity_comparison_Q1$sig[which(hill_phyto_diversity_comparison_Q1$D_q_norm >= 2)] <- "Overdisprsed"

#table()
hill_phyto_diversity_comparison_Q1 <- hill_phyto_diversity_comparison_Q1[order(hill_phyto_diversity_comparison_Q1$D_q_norm),]
hill_phyto_diversity_comparison_Q1$order <- 1:nrow(hill_phyto_diversity_comparison_Q1)

phyt_div_plot_Q1 <- ggbarplot(hill_phyto_diversity_comparison_Q1, x = "order", y = "D_q_norm",
          fill = "sig",               # change fill color by cyl
          color = "sig",            # Set bar border colors to white
          palette = "jco",            # jco journal color palett. see ?ggpar
          sort.val = "asc",          # Sort the value in dscending order
          sort.by.groups = FALSE,     # Don't sort inside each group
          x.text.angle = 90           # Rotate vertically x axis texts
          ) + xlab("") + ylab("Normalized Phytochemical Diversity") + theme(axis.title.x=element_blank(),legend.title = element_blank(),
        axis.text.x=element_blank(),axis.ticks.x=element_blank()) + scale_fill_manual(values = c("Underdispersed" = "red", "Random" = "Yellow","Overdisprsed" = "Blue")) + geom_hline(yintercept=2, linetype="dashed", color = "red") + geom_hline(yintercept= -2, linetype="dashed", color = "red")
        

ggsave(phyt_div_plot_Q1,path = here("./results/"),filename = "Figure_2_Phyt_div_plot_D_q_Q1.pdf",dpi = 600,width = 12)
phyt_div_plot_Q1
table(hill_phyto_diversity_comparison_Q1$sig)/nrow(hill_phyto_diversity_comparison_Q1)
#Q1 = 44 percent overdispered and 55 percent random

```

Figure 1B - bargraph Q=0

```{r}

hill_phyto_diversity_comparison_Q0$sig <- "Random"
hill_phyto_diversity_comparison_Q0$sig[which(hill_phyto_diversity_comparison_Q0$D_q_norm <= -2)] <- "Underdispersed"
hill_phyto_diversity_comparison_Q0$sig[which(hill_phyto_diversity_comparison_Q0$D_q_norm >= 2)] <- "Overdisprsed"

#table()
hill_phyto_diversity_comparison_Q0 <- hill_phyto_diversity_comparison_Q0[order(hill_phyto_diversity_comparison_Q0$D_q_norm),]
hill_phyto_diversity_comparison_Q0$order <- 1:nrow(hill_phyto_diversity_comparison_Q0)

phyt_div_plot_Q0 <- ggbarplot(hill_phyto_diversity_comparison_Q0, x = "order", y = "D_q_norm",
          fill = "sig",               # change fill color by cyl
          color = "sig",            # Set bar border colors to white
          palette = "jco",            # jco journal color palett. see ?ggpar
          sort.val = "asc",          # Sort the value in dscending order
          sort.by.groups = FALSE,     # Don't sort inside each group
          x.text.angle = 90           # Rotate vertically x axis texts
          ) + xlab("") + ylab("Normalized Phytochemical Diversity") + theme(axis.title.x=element_blank(),legend.title = element_blank(),
        axis.text.x=element_blank(),axis.ticks.x=element_blank()) + scale_fill_manual(values = c("Underdispersed" = "red", "Random" = "Yellow","Overdisprsed" = "Blue")) + geom_hline(yintercept=2, linetype="dashed", color = "red") + geom_hline(yintercept= -2, linetype="dashed", color = "red")
        

#ggsave(phyt_div_plot,path = here("./results/"),filename = "Figure_2_Phyt_div_plot_D_q.pdf",dpi = 600,width = 12)
phyt_div_plot_Q0
table(hill_phyto_diversity_comparison_Q0$sig)/nrow(hill_phyto_diversity_comparison_Q0)


ggsave(phyt_div_plot_Q0,path = here("./results/"),filename = "Figure_2_Phyt_div_plot_D_q_Q0.pdf",dpi = 600,width = 12)
table(hill_phyto_diversity_comparison_Q0$sig)/nrow(hill_phyto_diversity_comparison_Q0)
#Q0 = 04 percent overdispered,93% random and and 02 underisperesed

phyt_div_plot_Q0


```
